[Refactor] Move Tour Api test code to MainViewModel
