--- app/src/main/java/com/twoday/todaytrip/viewModel/RecordViewModel.kt
@@ -1,7 +1,0 @@
-package com.twoday.todaytrip.viewModel
-
-import androidx.lifecycle.ViewModel
-
-class RecordViewModel: ViewModel() {
-
-}
